+++ include/ECS/Motion.hpp
@@ -2,4 +2,4 @@
 
 #include "ECSManager.hpp"
 
-void Motion_System(ECS_Manager &world, float dt);
+void Motion_System(ECS_Manager &world, double dt);
