+++ include/ECS/components/Acceleration_comp.hpp
@@ -6,6 +6,6 @@ struct Acceleration_Component {
 
     int entity_id;
 
-    Eigen::Vector2f accel; // {x, y}
+    Eigen::Vector2d accel; // {x, y}
 
 };
