+++ include/ECS/components/Constraint_comp.hpp
@@ -7,9 +7,9 @@ struct Fixed_Rot_Component {
     int entity_id; 
 
     int constr_entity; 
-    Eigen::Vector2f fixed_point;
-    Eigen::Vector2f rel_body_pos;
-    float radius;
+    Eigen::Vector2d fixed_point;
+    Eigen::Vector2d rel_body_pos;
+    double radius;
 };
 
 struct Relative_Rot_Component {
@@ -20,9 +20,9 @@ struct Relative_Rot_Component {
     int constr_entity1;
     int constr_entity2;
     
-    Eigen::Vector2f rel_body_pos1;
-    Eigen::Vector2f rel_body_pos2; 
-    float radius;
+    Eigen::Vector2d rel_body_pos1;
+    Eigen::Vector2d rel_body_pos2; 
+    double radius;
 
 };
 
@@ -33,7 +33,7 @@ struct Linear_Component {
     int constr_entity;
     
     // Defines the line to be constrained to
-    float m;
-    float b;
+    double m;
+    double b;
 
 };
