+++ src/FlowFieldVisual.cpp
@@ -20,42 +20,42 @@
 // The (...)_X/Y functions would handle the transforms. 
 
 // I don't like this and it either needs to change or be made more clear which is which.
-static float screen2worldscale_X(int screen_x){
-    return screen_x * ((float)SCREEN_WIDTH_METERS/SCREEN_WIDTH_IN_PIXELS); 
+static double screen2worldscale_X(int screen_x){
+    return screen_x * ((double)SCREEN_WIDTH_METERS/SCREEN_WIDTH_IN_PIXELS); 
 };
 
-static float screen2worldscale_Y(int screen_y){
-    return -(screen_y - SCREEN_HEIGHT_IN_PIXELS)*((float)SCREEN_HEIGHT_METERS/SCREEN_HEIGHT_IN_PIXELS);
+static double screen2worldscale_Y(int screen_y){
+    return -(screen_y - SCREEN_HEIGHT_IN_PIXELS)*((double)SCREEN_HEIGHT_METERS/SCREEN_HEIGHT_IN_PIXELS);
 };
 
-static float screen2world_Y(int screen_y){
-    return  -((float)SCREEN_HEIGHT_METERS/SCREEN_HEIGHT_IN_PIXELS) * (screen_y - SCREEN_HEIGHT_IN_PIXELS);
+static double screen2world_Y(int screen_y){
+    return  -((double)SCREEN_HEIGHT_METERS/SCREEN_HEIGHT_IN_PIXELS) * (screen_y - SCREEN_HEIGHT_IN_PIXELS);
 };
 
 //Scale differences
-static float world2screenscale_X(float x){
-    return x * ((float)SCREEN_WIDTH_IN_PIXELS/SCREEN_WIDTH_METERS) + SCREEN_WIDTH_IN_PIXELS/2;  
+static double world2screenscale_X(double x){
+    return x * ((double)SCREEN_WIDTH_IN_PIXELS/SCREEN_WIDTH_METERS) + SCREEN_WIDTH_IN_PIXELS/2;  
 }
-static float world2screenscale_Y(float y){
-    return y * ((float)SCREEN_HEIGHT_IN_PIXELS/SCREEN_HEIGHT_METERS) + SCREEN_HEIGHT_IN_PIXELS/2;
+static double world2screenscale_Y(double y){
+    return y * ((double)SCREEN_HEIGHT_IN_PIXELS/SCREEN_HEIGHT_METERS) + SCREEN_HEIGHT_IN_PIXELS/2;
 }
 
 // Coord transform that assumes orthogonality for the transform
-static float world2screen_X(float x){
+static double world2screen_X(double x){
     return world2screenscale_X(x);
 }
-static float world2screen_Y(float y){
+static double world2screen_Y(double y){
     return -world2screenscale_Y(y) + SCREEN_HEIGHT_IN_PIXELS;
 }
 
 void FlowField_Visualization_System(ECS_Manager &world){
     
     
-    float max_len = 0.0;
+    double max_len = 0.0;
     for (auto it = world.get_component_begin<Vector_Component>(); 
               it < world.get_component_end<Vector_Component>(); it++){
-        Eigen::Vector2f vec = world.get_component<Vector_Component>(it->entity_id)->vec; 
-        float vec_mag = std::sqrt(vec(0)*vec(0) + vec(1)*vec(1));
+        Eigen::Vector2d vec = world.get_component<Vector_Component>(it->entity_id)->vec; 
+        double vec_mag = std::sqrt(vec(0)*vec(0) + vec(1)*vec(1));
 
         if (vec_mag > max_len){
             max_len = vec_mag;
@@ -65,12 +65,12 @@ void FlowField_Visualization_System(ECS_Manager &world){
     for (auto it = world.get_component_begin<Vector_Component>(); 
               it < world.get_component_end<Vector_Component>(); it++){
         // Take Physical coords. and convert to a location on the screen
-        Eigen::Vector2f physical_pos = world.get_component<Position_Component>(it->entity_id)->position;
+        Eigen::Vector2d physical_pos = world.get_component<Position_Component>(it->entity_id)->position;
         
-        Eigen::Vector2f vec = world.get_component<Vector_Component>(it->entity_id)->vec;
-        float vec_mag = std::sqrt(vec(0)*vec(0) + vec(1)*vec(1));
+        Eigen::Vector2d vec = world.get_component<Vector_Component>(it->entity_id)->vec;
+        double vec_mag = std::sqrt(vec(0)*vec(0) + vec(1)*vec(1));
 
-        float scale = vec_mag/max_len;
+        double scale = vec_mag/max_len;
         
         world.get_component<Render_Component>(it->entity_id)->x = world2screen_X(physical_pos(0));
         world.get_component<Render_Component>(it->entity_id)->y = world2screen_Y(physical_pos(1));
